pyproject.toml
[tool.poetry]
name = "nex-backend"
version = "0.1.0"
description = ""
authors = ["AnvanshSingh <anvansh.singh@esmagico.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
websockets = ">=13.0.0,<15.1.0"
pydantic = "^2.12.5"
python-dotenv = "^1.2.1"
loguru = "^0.7.3"
uvloop = "^0.22.1"
locust = "^2.43.1"
google-generativeai = "^0.8.6"
pymilvus = "^2.5.0"
httpx = ">=0.28.1,<1.0.0"
neo4j = "^6.1.0"
graphiti-core = {extras = ["google-genai"], version = "^0.26.3"}
firebase-admin = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
